Fix repeated config rewrites caused by light levels being set to 2^32-1
The value 255 gets converted to -1 which is 4294967295 in the uint8_fast_t
value which then gets written to the config as-is instead of writing -1.

Setting individual light levels also needs to convert -1 to 255 or they
will become 4294967295 too.

+++ src/config.h
@@ -126,7 +126,7 @@ private:
 
 class Config {
 public:
-	static constexpr int16_t LEVEL_NO_CHANGE = -1;
+	static constexpr int64_t LEVEL_NO_CHANGE = -1;
 
 	explicit Config(std::mutex &file_mutex, Network &network);
